edit title

+++ tools/list.sh
@@ -8,7 +8,9 @@
 # generate new post
 date=`date '+%Y%m%d'`
 logfile=$PATH_LOG/$date"_list.log"
-rm $logfile
+if [ -e $logfile ]; then
+    rm $logfile
+fi
 $PHP $PATH_TOOL/php/list.php --num=$MD_NUM >> $logfile
 
 cat $logfile
